requirements.txt
aiokeydb==0.1.38
anyio==3.6.2
async-lru==2.0.2
async-timeout==4.0.2
click==8.1.3
croniter==1.3.10
Deprecated==1.2.13
dill==0.3.6
fastapi==0.95.0
frozendict==2.3.6
h11==0.14.0
httptools==0.5.0
idna==3.4
lazyops==0.2.34
loguru==0.6.0
orjson==3.8.9
packaging==23.0
pydantic==1.10.7
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
PyYAML==6.0
redis==4.5.4
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
wrapt==1.15.0
